# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Position:
    properties:
      active:
        type: boolean
      archived:
        type: boolean
      company_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.PositionSet:
    properties:
      archived:
        type: boolean
      company_id:
        type: integer
      name:
        type: string
    type: object
...
  rest.sErr:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Описание API QuickOn
  title: QuickOn
  version: "0.1"
paths:
...
  /positions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Position'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.sErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.sErr'
      summary: Получение всех должностей
      tags:
      - position
    post:
      parameters:
      - description: Position Create
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.PositionSet'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Position'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.sErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.sErr'
      summary: Создание новой должности
      tags:
      - position
  /positions/{id}:
    get:
      parameters:
      - description: Position ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Position'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.sErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.sErr'
      summary: Получение всех должностей
      tags:
      - position
  /positions/update/{id}:
    patch:
      parameters:
      - description: Position ID
        in: path
        name: id
        required: true
        type: integer
      - description: Position info
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.PositionSet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Position'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.sErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.sErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.sErr'
      summary: Обновление данных о должности
      tags:
      - position
...
securityDefinitions:
  Bearer:
    description: you can get it on login page
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"